backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0

# Essential data dependencies (avoiding build issues)
yfinance==0.2.28
newsapi-python==0.2.7
vaderSentiment==3.3.2
requests==2.31.0
aiofiles==23.2.1
httpx==0.25.2

# Data processing (will install without compilation issues)
pandas
numpy
email-validator>=2.1          # or
pydantic[email]>=2.11
